requirements.txt
pydantic-settings==2.6.0
selenium==4.25.0
celery==5.4.0
motor==3.6.0
aiogram==3.13.1
flake8==7.1.1